requirements.txt
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0

click
dtpattern
pyanycsv
pandas
numpy
structlog
requests
cchardet
markupsafe
pyjuhelpers
jinja2
dateutil
